exporter/pyproject.toml
[tool.poetry]
name = "exporter"
version = "0.1.0"
description = "Container for exporting archives to long term preservation (az specific)"
authors = ["Per Buer <per.buer@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.7"
azure-storage-blob = "^12.3.2"
azure-servicebus = "^0.50.3"

[tool.poetry.dev-dependencies]
pylint = "^2.4.4"
python-dotenv = "^0.13.0"
autopep8 = "^1.5.3"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
